requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
PyYAML>=6.0.1
rich>=13.7.0
pandas>=2.2.0
pyarrow>=16.0.0
duckdb>=1.0.0

